+++ .trebired/bundler/config.ts
@@ -1,3 +1,4 @@
 export default {
+  forVersion: "5.1.3",
   prefix: "tbf",
 };